Created core structure of known database using Entity Framework 6 Code First design. In order for this to work on your system, you will need to go to Tools->NuGet Package Manager->Package manager console and type in Install-Package EntityFramework.
Removed most controllers as they will need to be created in a similar way to the CharterCarriersController.

Added the Data Access Layer folder that handles database processing - includes the database context class and a databasee initialization script that will run every time the application starts - this will be removed when going to production.

Added the required connection string to the Web.Config file - I named my SQL Server instance SQLServer2012, and this may need to be changed for your system. I also created the schema that the application will use and I named it charterSystemDB.

Added CRUD operations for the charter carrier table.

Added other stuff that I can't think of at the moment.