Added CharterAgreement functionality - add and edit - Agreements don't need to be deleted.
The details view will need to be reworked and the editing is pretty ugly for the adding of scheduled trips, but creating a charter agreement will generate all of the trips specified in the agreement.

CharterAgreements should also be fully compatible with the IValidatableObject implementation because the arrival dates and departure dates are bound on the edit and create form submits.